requirements.txt
pandas==2.2.2
numpy==1.26.4
pyyaml==6.0
scikit-learn==1.3.2
pytest==8.4.3